sim/Traditional family: new member w/ versions
This is a new allocator simulator that understands that memory may have
colors associated with it.  We assume that we can set the color per
granule to a value of our choosing and that we can read the color back.

The latter is used, beyond its obvious role in incrementing the
tag-as-version of memory, to turn every revocation pass into a
revocation of all maximal-version regions.  In this mode, this simulator
is not parametric in the number of revocation "windows" available to it.
This does seem to presume, however, the existence of a credible hardware
anti-store-behind concurrency story.

For the purposes of comparison, this new allocator implements many of
the knobs available on other allocators as well, including limited
revocation windows.

Moreover, interestingly, because of the immediacy of the revocation
provided by tags, this model could be instantiated without segregated
metadata.

While here, refactor some Base functionality into DLMallocish allocator.
The base class retains functionality used by either child instantiation.