Traditional family: introduce "PEND" state
This allows child allocator classes to hold on to memory even when it is
not allocated to the user proper.  The primary motivation is
experimentation with very large revocation (or MTE) granules; in these
cases, it may not be possible to express revocation (or invalidation) of
bytes, even if the allocator can operate with byte granularity.

This is kind of a fusion between the DLM and Cling state machines, if
one likens the granules to Cling buckets: once any part of a granule is
allocated, the whole thing is unreclaimable until all bytes are no
longer allocated.

Tenatively, I hope that our prior investigation into lifetime clustering
can help keep the intra-granule fragmentation low.